Refactored code so that handlers are attached to the buttons, the buttons inherit createjs.container, and the color is a simple createjs.shape that just gives color, making them easy to change out.
Also added speaker display while sound is playing as an indication something is happening if the computer is not hooked up to a speaker.

Also added mouseover/off to slightly change the color of the buttons to indicate that an event can occur if the mouse is clicked.

Changed name of CreateNewClockButton to CreateNewGameButton for a better description of the more generalized functionality